requirements.txt
ipython==8.9.0
PyYAML==6.0
schema==0.7.5